AgriSmart-main/backend/requirements-deploy.txt
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.5
gunicorn==21.2.0
joblib==1.5.3
numpy==2.4.2
pandas==2.3.3
pillow==11.3.0
scikit-learn==1.8.0
scipy==1.16.3
torch==2.10.0
torchvision==0.25.0
SQLAlchemy==2.0.45
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
